refactor(engine): split opencode command surface
