+++ docs/architecture/large-file-baseline.md
@@ -1,11 +1,10 @@
 # Large File Hard-Debt Baseline
 
-- Generated at: 2026-04-22T17:37:49.723Z
+- Generated at: 2026-04-22T17:54:39.252Z
 - Scope: fail
 - Policy version: 2026-04-22.policy-v2
-- Count: 1
+- Count: 0
 
 | File | Lines | Type | Policy | Priority | Warn | Fail | Severity | Status | Baseline | Delta |
 |---|---:|---|---|---|---:|---:|---|---|---:|---:|
-| `src-tauri/src/engine/commands.rs` | 2976 | rust | bridge-runtime-critical | P0 | 2200 | 2600 | fail | captured |  |  |
